+++ tests/Aspire.Npgsql.EntityFrameworkCore.PostgreSQL.Tests/AspireEFPostgreSqlExtensionsTests.cs
@@ -125,103 +125,22 @@ public class AspireEFPostgreSqlExtensionsTests
     }
 
     [Fact]
-    public void EnrichCanConfigureDbContextOptions()
+    public void AddNpgsqlCanConfigureDbContextOptionsWithoutRetry()
     {
         var builder = Host.CreateEmptyApplicationBuilder(null);
         builder.Configuration.AddInMemoryCollection([
-            new KeyValuePair<string, string?>("Aspire:Npgsql:EntityFrameworkCore:PostgreSQL:Retry", "true")
-        ]);
-
-        builder.Services.AddDbContextPool<TestDbContext>(optionsBuilder =>
-        {
-            ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString, npgsqlBuilder =>
-            {
-                npgsqlBuilder.CommandTimeout(123);
-            });
-        });
-
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        var host = builder.Build();
-        var context = host.Services.GetRequiredService<TestDbContext>();
-
-#pragma warning disable EF1001 // Internal EF Core API usage.
-
-        var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
-        Assert.NotNull(extension);
-
-        // ensure the command timeout was respected
-        Assert.Equal(123, extension.CommandTimeout);
-
-        // ensure the max retry count from config was respected
-        Assert.NotNull(extension.ExecutionStrategyFactory);
-        var executionStrategy = extension.ExecutionStrategyFactory(new ExecutionStrategyDependencies(new CurrentDbContext(context), context.Options, null!));
-        var retryStrategy = Assert.IsType<NpgsqlRetryingExecutionStrategy>(executionStrategy);
-        Assert.Equal(new WorkaroundToReadProtectedField(context).MaxRetryCount, retryStrategy.MaxRetryCount);
-
-#pragma warning restore EF1001 // Internal EF Core API usage.
-    }
-
-    [Fact]
-    public void EnrichEnablesRetryByDefault()
-    {
-        var builder = Host.CreateEmptyApplicationBuilder(null);
-
-        builder.Services.AddDbContextPool<TestDbContext>(optionsBuilder =>
-        {
-            ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString);
-        });
-
-        var oldOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TestDbContext>));
-        Assert.NotNull(oldOptionsDescriptor);
-
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        var host = builder.Build();
-        var context = host.Services.GetRequiredService<TestDbContext>();
-
-#pragma warning disable EF1001 // Internal EF Core API usage.
-
-        var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
-        Assert.NotNull(extension);
-
-        // ensure the max retry count from config was respected
-        Assert.NotNull(extension.ExecutionStrategyFactory);
-        var executionStrategy = extension.ExecutionStrategyFactory(new ExecutionStrategyDependencies(new CurrentDbContext(context), context.Options, null!));
-        var retryStrategy = Assert.IsType<NpgsqlRetryingExecutionStrategy>(executionStrategy);
-        Assert.Equal(new WorkaroundToReadProtectedField(context).MaxRetryCount, retryStrategy.MaxRetryCount);
-
-#pragma warning restore EF1001 // Internal EF Core API usage.
-    }
-
-    [Fact]
-    public void EnrichPreservesDefaultWhenMaxRetryCountNotSet()
-    {
-        var builder = Host.CreateEmptyApplicationBuilder(null);
-        builder.Configuration.AddInMemoryCollection([
+            new KeyValuePair<string, string?>("ConnectionStrings:npgsql", ConnectionString),
             new KeyValuePair<string, string?>("Aspire:Npgsql:EntityFrameworkCore:PostgreSQL:Retry", "false")
         ]);
 
-        builder.Services.AddDbContextPool<TestDbContext>(optionsBuilder =>
+        builder.AddNpgsqlDbContext<TestDbContext>("npgsql", configureDbContextOptions: optionsBuilder =>
         {
             ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString, npgsqlBuilder =>
+            optionsBuilder.UseNpgsql(npgsqlBuilder =>
             {
-                npgsqlBuilder.EnableRetryOnFailure(456);
+                npgsqlBuilder.CommandTimeout(123);
             });
         });
-
-        var oldOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TestDbContext>));
-        Assert.NotNull(oldOptionsDescriptor);
-
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        // The service descriptor should not be affected with MaxRetryCount set to 0
-        var optionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TestDbContext>));
-        Assert.NotNull(optionsDescriptor);
-        Assert.Same(oldOptionsDescriptor, optionsDescriptor);
 
         var host = builder.Build();
         var context = host.Services.GetRequiredService<TestDbContext>();
@@ -231,113 +150,16 @@ public class AspireEFPostgreSqlExtensionsTests
         var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
         Assert.NotNull(extension);
 
-        // ensure the max retry count from config was respected
-        Assert.NotNull(extension.ExecutionStrategyFactory);
-        var executionStrategy = extension.ExecutionStrategyFactory(new ExecutionStrategyDependencies(new CurrentDbContext(context), context.Options, null!));
-        var retryStrategy = Assert.IsType<NpgsqlRetryingExecutionStrategy>(executionStrategy);
-        Assert.Equal(456, retryStrategy.MaxRetryCount);
-
-#pragma warning restore EF1001 // Internal EF Core API usage.
-    }
-
-    [Fact]
-    public void EnrichOverridesCustomRetryIfNotDisabled()
-    {
-        var builder = Host.CreateEmptyApplicationBuilder(null);
-        builder.Configuration.AddInMemoryCollection([
-            new KeyValuePair<string, string?>("Aspire:Npgsql:EntityFrameworkCore:PostgreSQL:Retry", "true")
-        ]);
-
-        builder.Services.AddDbContextPool<TestDbContext>(optionsBuilder =>
-        {
-            ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString, npgsqlBuilder =>
-            {
-                npgsqlBuilder.EnableRetryOnFailure(456);
-            });
-        });
-
-        var oldOptionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TestDbContext>));
-        Assert.NotNull(oldOptionsDescriptor);
-
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        var host = builder.Build();
-        var context = host.Services.GetRequiredService<TestDbContext>();
-
-#pragma warning disable EF1001 // Internal EF Core API usage.
-
-        var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
-        Assert.NotNull(extension);
-
-        // ensure the max retry count from config was respected
-        Assert.NotNull(extension.ExecutionStrategyFactory);
-        var executionStrategy = extension.ExecutionStrategyFactory(new ExecutionStrategyDependencies(new CurrentDbContext(context), context.Options, null!));
-        var retryStrategy = Assert.IsType<NpgsqlRetryingExecutionStrategy>(executionStrategy);
-        Assert.Equal(new WorkaroundToReadProtectedField(context).MaxRetryCount, retryStrategy.MaxRetryCount);
-
-#pragma warning restore EF1001 // Internal EF Core API usage.
-    }
-
-    [Fact]
-    public void EnrichSupportServiceType()
-    {
-        var builder = Host.CreateEmptyApplicationBuilder(null);
-
-        builder.Services.AddDbContextPool<ITestDbContext, TestDbContext>(optionsBuilder =>
-        {
-            ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString, npgsqlBuilder =>
-            {
-                npgsqlBuilder.CommandTimeout(123);
-            });
-        });
-
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        var host = builder.Build();
-        var context = host.Services.GetRequiredService<ITestDbContext>() as TestDbContext;
-        Assert.NotNull(context);
-
-#pragma warning disable EF1001 // Internal EF Core API usage.
-        var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
-        Assert.NotNull(extension);
-
         // ensure the command timeout was respected
         Assert.Equal(123, extension.CommandTimeout);
-#pragma warning restore EF1001 // Internal EF Core API usage.
-    }
 
-    [Fact]
-    public void EnrichSupportCustomOptionsLifetime()
-    {
-        var builder = Host.CreateEmptyApplicationBuilder(null);
+        // ensure the connection string from config was respected
+        var actualConnectionString = context.Database.GetDbConnection().ConnectionString;
+        Assert.Equal(ConnectionString, actualConnectionString);
 
-        builder.Services.AddDbContext<ITestDbContext, TestDbContext>(optionsBuilder =>
-        {
-            ConfigureDbContextOptionsBuilderForTesting(optionsBuilder);
-            optionsBuilder.UseNpgsql(ConnectionString, npgsqlBuilder =>
-            {
-                npgsqlBuilder.CommandTimeout(123);
-            });
-        }, contextLifetime: ServiceLifetime.Singleton);
+        // ensure no retry strategy was registered
+        Assert.Null(extension.ExecutionStrategyFactory);
 
-        builder.EnrichNpgsqlDbContext<TestDbContext>();
-
-        var optionsDescriptor = builder.Services.FirstOrDefault(sd => sd.ServiceType == typeof(DbContextOptions<TestDbContext>));
-        Assert.NotNull(optionsDescriptor);
-        Assert.Equal(ServiceLifetime.Singleton, optionsDescriptor.Lifetime);
-
-        var host = builder.Build();
-        var context = host.Services.GetRequiredService<ITestDbContext>() as TestDbContext;
-        Assert.NotNull(context);
-
-#pragma warning disable EF1001 // Internal EF Core API usage.
-        var extension = context.Options.FindExtension<NpgsqlOptionsExtension>();
-        Assert.NotNull(extension);
-
-        // ensure the command timeout was respected
-        Assert.Equal(123, extension.CommandTimeout);
 #pragma warning restore EF1001 // Internal EF Core API usage.
     }
 
@@ -349,5 +171,4 @@ public class AspireEFPostgreSqlExtensionsTests
 
         public int RetryCount => base.MaxRetryCount;
     }
-
 }
